requirements.txt
ollama
streamlit
pandas
loguru
python-dotenv
gradio